+++ src/Token.php
@@ -4,15 +4,21 @@ declare(strict_types=1);
 
 namespace Bakame\Http\StructuredFields;
 
+use Stringable;
 use function preg_match;
 
 final class Token implements StructuredField
 {
-    public function __construct(private string $value)
+    private function __construct(private string $value)
     {
         if (1 !== preg_match("/^([a-z*][a-z0-9:\/\!\#\$%&'\*\+\-\.\^_`\|~]*)$/i", $this->value)) {
             throw new SyntaxError('Invalid characters in token');
         }
+    }
+
+    public static function fromString(string|Stringable $value): self
+    {
+        return new self((string) $value);
     }
 
     /**